@@ -14,30 +14,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.0",
-    "@angular/cdk": "^19.1.0",
-    "@angular/common": "^19.0.0",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
-    "@angular/forms": "^19.0.0",
-    "@angular/material": "^19.1.0",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/platform-browser-dynamic": "^19.0.0",
-    "@angular/platform-server": "^19.0.0",
-    "@angular/router": "^19.0.0",
+    "@angular/animations": "^19.1.6",
+    "@angular/cdk": "^19.1.4",
+    "@angular/common": "^19.1.6",
+    "@angular/compiler": "^19.1.6",
+    "@angular/core": "^19.1.6",
+    "@angular/forms": "^19.1.6",
+    "@angular/material": "^19.1.4",
+    "@angular/platform-browser": "^19.1.6",
+    "@angular/platform-browser-dynamic": "^19.1.6",
+    "@angular/platform-server": "^19.1.6",
+    "@angular/router": "^19.1.6",
     "@angular/ssr": "^19.0.7",
     "@ngx-translate/core": "^16.0.4",
     "@ngx-translate/http-loader": "^16.0.1",
     "express": "^4.18.2",
     "idb": "^8.0.1",
     "rxjs": "~7.8.0",
+    "rxjs-interop": "^2.0.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^19.1.7",
-    "@angular/cli": "^19.0.7",
-    "@angular/compiler-cli": "^19.0.0",
+    "@angular/cli": "^19.1.7",
+    "@angular/compiler-cli": "^19.1.6",
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
