@@ -8,11 +8,17 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatCard } from '@angular/material/card';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { BaseModel } from '../../../shared/models/api/base.model';
-import { DetailComponent } from '../../../shared/components/detail/detail.component';
 import { Project } from '../../../shared/models/project.model';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ProjectsService } from '../../../core/services/api/projects.service';
 import { MatDivider } from '@angular/material/divider';
+import { AbstractInformationComponent } from '../../../shared/components/abstract-information/abstract-information.component';
+import { MatDialog } from '@angular/material/dialog';
+import { ProjectFormComponent } from '../project-form/project-form.component';
+import { tap } from 'rxjs';
+import { WorkFormComponent } from '../../works/work-form/work-form.component';
+import { MatIconButton } from '@angular/material/button';
+import { MatIcon } from '@angular/material/icon';
 
 @Component({
   selector: 'app-project-info',
@@ -28,22 +34,23 @@ import { MatDivider } from '@angular/material/divider';
     MatDatepickerModule,
     NgFor,
     MatDivider,
+    MatIcon,
+    MatIconButton,
   ],
   templateUrl: './project-info.component.html',
   styleUrl: './project-info.component.scss',
 })
-export class ProjectInfoComponent extends DetailComponent<Project> {
-  // TODO: remove after DetailComponent split
-  override newData(): Project {
-    throw new Error('Method not implemented.');
-  }
-
-  // TODO: remove after DetailComponent split
-  override getFormattedData(): Project {
-    throw new Error('Method not implemented.');
+export class ProjectInfoComponent extends AbstractInformationComponent<Project> {
+  override edit(data: Project): void {
+    this.matDialog
+      .open(ProjectFormComponent, { data })
+      .beforeClosed()
+      .pipe(tap(() => this.refresh.next(data.id as string)))
+      .subscribe();
   }
 
   constructor(
+    public matDialog: MatDialog,
     activatedRoute: ActivatedRoute,
     projectsService: ProjectsService,
     location: Location,
@@ -54,5 +61,13 @@ export class ProjectInfoComponent extends DetailComponent<Project> {
 
   onRowClick(row: BaseModel): void {
     this.router.navigate(['works', row.id]);
+  }
+
+  openWorkCreationDialog(dataId: string) {
+    this.matDialog
+      .open(WorkFormComponent, { data: { parentProject: { id: dataId } } })
+      .beforeClosed()
+      .pipe(tap(() => this.refresh.next(dataId)))
+      .subscribe();
   }
 }
