@@ -1,4 +1,5 @@
 import { Work } from '../shared/models/work.model';
 import { Project } from '../shared/models/project.model';
+import { WorkType } from '../shared/models/work-type.model';
 
-export type BaseModelImpls = Project | Work;
+export type BaseModelImpls = Project | Work | WorkType;
